Integrate Payment functionality
Add animation for PaymentScreen.
Add implementation in PaymentDataSource to
manage data related to PaymentScreen.
Add PaymentViewModel implementation for
managing UI states.
Lifting multiple state up to the PaymentScreen.
Add convertTimeStamp in TimeStampSource.
Update PaymentHistory to align to the design
in PaymentViewModel.
Update PaymentHistoryDao to add order and
remove unnecessary query.
Rename PaymentState in HistoryScreen to
PaymentHistoryState.
Update unit test and remove not needed
tests.

Test: Unit test is passing
Test: Can assemble the project
Test: No problem when running spotlessCheck
Test: Can install and run the application

+++ app/src/test/java/com/example/tipjar/model/repository/PaymentRepositoryTest.kt
@@ -20,7 +20,7 @@ open class PaymentRepositoryTest {
     internal val defaultPaymentHistory =
         PaymentHistory(
             timestamp = "",
-            amount = defaultPayment.amount,
+            amount = defaultPayment.amount.toFloat(),
             tip = defaultPayment.totalTip,
             image = "",
         )
